1) Refactor DTWAIN_Acquire... functions  2) Fix DIB issue with DTWAIN_AcquireClipboard

+++ source/cpp/pdflib32/base64encode.cpp
@@ -50,5 +50,5 @@ int Base64Encode(const uint8_t* inData, std::string& out, size_t len)
             out.push_back('=');
     }
 
-    return out.size();
+    return static_cast<int>(out.size());
 }
